Python_files/requirements.txt
langchain==0.1.0
langchain_community==0.0.11
langchain_core==0.1.8
mysql_connector==8.2.0
pyTelegramBotAPI==4.14.1
openai==1.6.1